combine step function utils into single file
